reworked enemy turn function
in battle manager I reworked the enmy turn function to use the modulos in order to determine what enemy in the enemy list's turn it is. Also implemented enemy attack function into it. As well as added End Turn and Attack Finished variables to the Enemy header file for the cpp to call.